pyproject.toml
[tool.poetry]
name = "YupiAI"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-decouple = "^3.8"
loguru = "^0.7.2"
mistralai = "^1.2.3"
duckduckgo-search = "^6.3.7"
gradio-client = "^1.5.0"
pypdf2 = "^3.0.1"
pillow = "^11.0.0"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.26.1"
trafilatura = "^2.0.0"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
